Added token expiration logic.

+++ fintrack-backend/app/config.py
@@ -5,7 +5,7 @@ class Settings(BaseSettings):
     DATABASE_URL: str
     SECRET_KEY: str
     ALGORITHM: str
-    ACCESS_TOKEN_EXPIRE_MINUTES: int
+    ACCESS_TOKEN_EXPIRE_MINUTES: float
     model_config = SettingsConfigDict(env_file=".env")
